extract matrix and species interfaces into dedicated module
